# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Volunteer Service Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all calendar events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
...
  /v1/events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found

    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an existing event (admin only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
  /v1/events/{eventId}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register for an upcoming event (participant)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/events/{eventId}/checkin:
    post:
      operationId: checkInToEvent
      tags: [events]
      summary: Check in to a day-of event (participant)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResponse"
        "400":
          description: Outside time window or not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
        - status
        - message
...
    Event:
      type: object
      required: [eventId, title, description, location, eventDate, startTime, endTime, hoursValue, maxCapacity, supervisorId, registrationCount]
      properties:
        eventId:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        eventDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        hoursValue:
          type: number
        maxCapacity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        supervisorId:
          type: string
        supervisorName:
          type: ["string", "null"]
        supervisorEmail:
          type: ["string", "null"]
        registrationCount:
          type: integer
        myRegistrationStatus:
          type: ["string", "null"]
        organizationId:
          type: ["string", "null"]
        organizationName:
          type: ["string", "null"]
        eligibleForMe:
          type: boolean
    EventInput:
      type: object
      required: [title, description, location, eventDate, startTime, endTime, maxCapacity, supervisorId]
      properties:
        title:
          type: string
          maxLength: 150
        description:
          type: string
        location:
          type: string
        eventDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        maxCapacity:
          type: integer
        supervisorId:
          type: string
        imageUrl:
          type: string
        organizationId:
          type: string
    EventUpdateInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 150
        description:
          type: string
        location:
          type: string
        eventDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        maxCapacity:
          type: integer
        supervisorId:
          type: string
        imageUrl:
          type: ["string", "null"]
        organizationId:
          type: ["string", "null"]
    EventRegistration:
      type: object
      required: [registrationId, eventId, userId, status, registeredAt]
      properties:
        registrationId:
          type: string
        eventId:
          type: string
        userId:
          type: string
        status:
          type: string
        registeredAt:
          type: string
        eventTitle:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        hoursValue:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        supervisorName:
          type: ["string", "null"]
        supervisorEmail:
          type: ["string", "null"]
    CheckInResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
        message:
          type: string
...